README: Update line-length error to 100 chars
Today I was reminded again that the current line length limit isn't 80
chars, but 100 chars, so I wanted to submit a PR to fix that in the
original repo, only to find someone else already did (PR 25).
And that's when I found this repo.
While the rule was updated, the README wasn't, so do that now.

Note that the current upstream 'Linux kernel coding style' does NOT
mention the 100 char limit, but only mentions the preferred max length
of 80.

Signed-off-by: Diederik de Haas <didi.debian@cknow.org>
Link: https://github.com/vivien/vim-linux-coding-style/pull/25
Link: https://git.kernel.org/linus/bdc48fa11e46f867ea4d75fa59ee87a7f48be144