README: Fix several URLs
The URL of this repo is (now?) ``kernel-coding-style`` and no longer
``vim-linux-coding-style``, so update that URL.

While `www.vim.org` is a valid URL, it leads to a 404 on Github, so
prefix that with https, so that it works correctly.
Update the URL to the vim scripts page to use https.

While at it, fix a minor typo.

Signed-off-by: Diederik de Haas <didi.debian@cknow.org>