Add syntax highlighting for guards
Define a new LinuxGuard syntax group and assign "guard", "scoped_guard",
and "scoped_cond_guard" to it. Link the new group with cConditional
group so that the same highlighting that is used for "if" and "while"
applied to the guard statements.

Add "return_ptr" to the list of "cStatements" to match regular "return".

Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>